Remove Confusing Wikipedia Link
Remove the link to Wikipedia for mailing list. It's pretty clear what a mailing list is and this link looks like it will helpfully take one to the mailing lists, when in fact it takes you further away from the mailing lists, which are at the bottom of the page after all the guidelines language. I can't be the only person who has clicked this link and been surprised to go to Wikipedia. This is the only time this week I'm going to lament the lack of a nice Google Analytics chart showing how many other users do the same thing, it would be fascinating to see.